requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
langchain_groq
duckduckgo-search
mysql-connector-python
SQLALchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
blinker